requirements.txt
alembic==1.10.4
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
autoflake==2.2.0
bcrypt==4.0.1
billiard==4.1.0
black==23.7.0
celery==5.3.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dnspython==2.3.0
docopt==0.6.2
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.2
fastapi==0.95.0
fastapi-cache2==0.2.1
flake8==6.0.0
flower==2.0.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
humanize==4.7.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.1
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
orjson==3.8.9
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
pendulum==2.1.2
Pillow==10.0.0
platformdirs==3.9.1
pluggy==1.2.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
pyasn1==0.5.0
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
pytest==7.4.0
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
redis==4.6.0
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqladmin==0.13.0
SQLAlchemy==2.0.13
starlette==0.26.1
tomli==2.0.1
tornado==6.3.2
typing_extensions==4.5.0
tzdata==2023.3
ujson==5.7.0
urllib3==2.0.3
uvicorn==0.21.1
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.1
WTForms==3.0.1
yarg==0.1.9
